+++ sdk/resourcemanager/redis/armredis/firewallrules_client.go
@@ -233,7 +233,6 @@ func (client *FirewallRulesClient) getHandleResponse(resp *http.Response) (Firew
 }
 
 // NewListPager - Gets all firewall rules in the specified redis cache.
-// If the operation fails it returns an *azcore.ResponseError type.
 // Generated from API version 2022-05-01
 // resourceGroupName - The name of the resource group.
 // cacheName - The name of the Redis cache.